ES6 Conversion
Convert closure library to use es6 import/export instead of goog.provide/goog.module.
The code appears to pass closure compiler type checking and the majority of tests are passing.
Closure Compiler needs an update to be able to correctly use this with advanced compilation, see https://github.com/google/closure-compiler/pull/4180

Of 632 test pages, 601 passed, 31 failed.
Of 10074 test functions, 9950 passed and 124 failed.

The majority of the remainging failures are cases where my technique for stubbing in es6 modules breaks down. When stubbing an es6 export, the stub is visible in anything importing the file, but not in the original file. Also, it leads to changes in an exported variable only being visible inside that file.

Other than that:
 - tests for debug loader and goog.modules aren't working (3 files, 4 testcases)
 - 1 failure in i18n/externs_test - testPluralOptions don't handle newer version of Chrome's Intl.PluralRules
 - the moduleloader doesn't work with es6 modules (23 failures)
 - portnetwork tests timeout for unknown reasons (1 failure)
 - 9 multitestrunner_test failures for unknown reasons. However, the multitestrunner seems to work fine for running all the tests.
 - 36 failures in crosspagechannel_test for unknown reasons
 - 3 failures in portoperator_test appear to be bad tests (comparison of mock method arguments is too strict)