Stop using obsolete closure primitives
Closure compiler looks for these based on the resolved name, which no longer works in es6 modules. So we inline what the compiler would do.
 - object.create -> object literal
 - object.createSet -> object literal with `true` values
 - reflect.objectProperty -> JSCompiler_renameProperty
 - reflect.object => multiple JSCompiler_renameProperty
 - add @pureOrBreakMyCode to reflect.cache calls

There's no documentation for @pureOrBreakMyCode, but Steve Hicks from Google said that's the correct solution and this seems to be the syntax based on https://github.com/google/closure-compiler/issues/4178
It seems like adding this to reflect.cache itself would work if the call is inlined. However I don't know if that always happens, so I also added it at the callsites.